Lots of cleanup for sequences, vectors, and strings. Particularly strings. Need to implement most of the string functions next and tidy up adjust-array.

+++ jcl-compiler/src/main/java/jcl/compiler/function/expanders/CompiledMacroFunctionExpander.java
@@ -86,7 +86,7 @@ public abstract class CompiledMacroFunctionExpander<O extends LispStruct> extend
 			functionParametersToBind.add(wholeParameterBinding);
 		}
 
-		final int numberOfArguments = Math.toIntExact(destructuredForm.length());
+		final int numberOfArguments = Math.toIntExact(destructuredForm.length().longValue());
 		final Iterator<LispStruct> functionArgumentsIterator = destructuredForm.iterator();
 
 		final String functionClassName = getClass().getSimpleName();
